更新TypeId类型为uint64_t，重构锁类以使用boost::noncopyable，添加Mutex和CountDownLatch的try_lock方法，优化锁的使用，删除冗余的非拷贝类

+++ bbt/core/reflex/TypeInfo.hpp
@@ -3,7 +3,7 @@
 namespace bbt::core::reflex
 {
 
-typedef int TypeId;
+typedef uint64_t TypeId;
 
 class Meta
 {
